:sparkles: Make archiver wait until all resources are loaded.
This replaces the 1s wait inside of Watcher with waiting on a set of two
promises. The first is a global network timer. The global network timer
will reject if it expires and there are still network requests pending.

The second promise is a network idle promise that will resolve if
playwright detects that the network has not been used in some amount of
time.

Fixes CAP-1082.