move most of shared package deps to devDependencies
Regular dependencies of the shared package will be bundled up into the external
pw/cy packages code when they are built, unless they are also dependencies of
those packages.

This moves move of shared's dependencies to devDependencies so they are
bundled up with its code rather than the pw/cy code. Either works, but this
feels a bit cleaner.

The couple of packages that haven't been moved are also dependencies of the external
pw/cy packages, so they won't get bundled up in those packages. The
`@segment/analytics-node` package does not need to be a dependency of those,
but it causes compilation warnings when it is bundled so this is to avoid that.